updating to long long timestamp_

+++ src/measurement_package.h
@@ -6,7 +6,7 @@
 class MeasurementPackage {
 public:
 
-  long timestamp_;
+  long long timestamp_;
 
   enum SensorType{
     BOTH,
